% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{as_sparse}
\alias{as_sparse}
\title{Transforms matrix to dgTMatrix}
\usage{
as_sparse(x)
}
\arguments{
\item{x}{Matrix or vector. If this is a vector, it is formatted into
a single-column matrix.}
}
\value{
Sparse matrix of class dgTMatrix
}
\description{
Transforms matrix to dgTMatrix
}
