% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{bdiag_check}
\alias{bdiag_check}
\title{Create block diagonal matrix (safe version)}
\usage{
bdiag_check(...)
}
\arguments{
\item{...}{Matrix or list of matrices (only the first argument is used)}
}
\value{
Block diagonal matrix
}
\description{
This version of bdiag checks whether the matrices passed as
arguments are NULL. This avoids errors that would arise if
using bdiag directly.
}
