% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{check_contiguous}
\alias{check_contiguous}
\title{Check values in vector are contiguous}
\usage{
check_contiguous(x)
}
\arguments{
\item{x}{Vector of values (can be numeric, character, factor)}
}
\value{
Logical: are values contiguous?
}
\description{
Check values in vector are contiguous
}
