% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{cov_grid}
\alias{cov_grid}
\title{Grid of covariates}
\usage{
cov_grid(var, data, covs = NULL, formulas)
}
\arguments{
\item{var}{Name of variable}

\item{data}{Data frame containing the covariates}

\item{covs}{Optional data frame with a single row and one column
for each covariate, giving the values that should be used. If this is
not specified, the mean value is used for numeric variables, and the
first level for factor variables.}

\item{formulas}{List of formulas used in the model}
}
\value{
Data frame of covariates, with 'var' defined over a grid,
and other covariates fixed to their mean (numeric) or first level
(factor).
}
\description{
Grid of covariates
}
