% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{dvonmises}
\alias{dvonmises}
\title{Density function of von Mises distribution}
\usage{
dvonmises(x, mu, kappa, log = FALSE)
}
\arguments{
\item{x}{Angle}

\item{mu}{Mean parameter}

\item{kappa}{Concentration parameter}

\item{log}{Should log-density be returned?}
}
\value{
Von Mises density
}
\description{
Density function of von Mises distribution
}
