% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{dwrpcauchy}
\alias{dwrpcauchy}
\title{Density function of wrapped Cauchy distribution}
\usage{
dwrpcauchy(x, mu, rho, log = FALSE)
}
\arguments{
\item{x}{Angle}

\item{mu}{Mean parameter}

\item{rho}{Concentration parameter}

\item{log}{Should log-density be returned?}
}
\value{
Wrapped Cauchy density
}
\description{
Density function of wrapped Cauchy distribution
}
