% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{find_re}
\alias{find_re}
\title{Find s(, bs = "re") terms in formula}
\usage{
find_re(form)
}
\arguments{
\item{form}{Model formula}
}
\value{
Vector of names of variables for which a random
effect term is included in the model.
}
\description{
This function is used to identify the variables "x" which are 
included as s(x, bs = "re") in the formula, in particular to
check that they are factors.
}
