% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{gdeterminant}
\alias{gdeterminant}
\title{Generalized matrix determinant}
\usage{
gdeterminant(x, eps = 1e-10, log = TRUE)
}
\arguments{
\item{x}{Numeric matrix}

\item{eps}{Threshold below which eigenvalues are ignored (default: 1e-10)}

\item{log}{Logical: should the log-determinant be returned?}
}
\value{
Generalized determinant of input matrix
}
\description{
Generalized determinant = product of non-zero eigenvalues 
(see e.g., Wood 2017). Used for (log)determinant of penalty matrices,
required in log-likelihood function.
}
