% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{HMMData}
\alias{HMMData}
\title{R6 class for HMM data}
\description{
R6 class for HMM data

R6 class for HMM data
}
\details{
Contains a data frame containing observations for the response variables,
as well as the covariates.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{HMMData$new()}}
\item \href{#method-data}{\code{HMMData$data()}}
\item \href{#method-ID}{\code{HMMData$ID()}}
\item \href{#method-clone}{\code{HMMData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create new HMMData object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HMMData$new(data, interval = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A data frame with at least one column of observations. A column
called "time" can be given for the time of each observation, otherwise
it is assumed that observations are regular in time. Other columns should
include "ID" (time series ID) if necessary, and any covariates needed in
the model.}

\item{\code{interval}}{Time interval of observation. Optional.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new HMMData object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data"></a>}}
\if{latex}{\out{\hypertarget{method-data}{}}}
\subsection{Method \code{data()}}{
Data frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HMMData$data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ID"></a>}}
\if{latex}{\out{\hypertarget{method-ID}{}}}
\subsection{Method \code{ID()}}{
Get IDs
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HMMData$ID()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Vector of time series IDs. Defaults to a vector
of 1s if there is no ID columns in the input data set.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HMMData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
