% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\docType{data}
\name{hmmTMB_cols}
\alias{hmmTMB_cols}
\title{hmmTMB colour palette}
\format{
An object of class \code{character} of length 6.
}
\usage{
hmmTMB_cols
}
\description{
hmmTMB colour palette
}
\keyword{datasets}
