% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{is_whole_number}
\alias{is_whole_number}
\title{Check if number of whole number}
\usage{
is_whole_number(x, tol = 1e-10)
}
\arguments{
\item{x}{number to check or vector of numbers}

\item{tol}{how far away from whole number is ok?}
}
\value{
TRUE if it is a whole number within tolerance
}
\description{
Check if number of whole number
}
