% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{load_distnames}
\alias{load_distnames}
\title{Load table of distribution names in hmmTMB}
\usage{
load_distnames()
}
\value{
loads table into variable called "distnames"
}
\description{
Load table of distribution names in hmmTMB
}
