% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik.HMM}
\alias{logLik.HMM}
\title{logLik function for HMM class}
\usage{
\method{logLik}{HMM}(object, ...)
}
\arguments{
\item{object}{the HMM model object}
}
\description{
logLik function for HMM class
}
