% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{make_cov}
\alias{make_cov}
\title{Make covariance matrix from standard deviations and correlations}
\usage{
make_cov(sds, corr)
}
\arguments{
\item{sds}{Vector of standard deviations}

\item{corr}{Vector of correlations (must be of length m*(m-1)/2 if
sds is of length m)}
}
\value{
An m by m covariance matrix
}
\description{
Make covariance matrix from standard deviations and correlations
}
