% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_formulas.R
\name{make_formulas}
\alias{make_formulas}
\title{Process formulas and store in nested list}
\usage{
make_formulas(input_forms, n_states)
}
\arguments{
\item{input_forms}{Nested list of formulas, with two levels: observed variable, 
and parameter of the observation distribution. The formulas can contain 
state-specific terms, e.g. "~ state1(x1) + x2".}

\item{n_states}{Number of states}
}
\value{
Nested list of formulas, with three levels: observed variable,
parameter of the observation distribution, and state.
}
\description{
Process formulas and store in nested list
}
