% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mat.R
\name{make_mat_hid}
\alias{make_mat_hid}
\title{Create model matrices (transition probabilities)}
\usage{
make_mat_hid(formulas, data)
}
\arguments{
\item{formulas}{List of formulas}

\item{data}{Data frame including covariates}
}
\value{
A list of
\itemize{
  \item X_fe Design matrix for fixed effects
  \item X_re Design matrix for random effects
  \item S Smoothness matrix
  \item ncol_re Number of columns of X_re and S for each random effect
}
}
\description{
Create model matrices (transition probabilities)
}
