% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mat.R
\name{make_mat_obs}
\alias{make_mat_obs}
\title{Create model matrices (observation model)}
\usage{
make_mat_obs(formulas, data, new_data = NULL)
}
\arguments{
\item{formulas}{Nested list of formulas, with three levels: (1) variable,
(2) parameter, and (3) state.}

\item{data}{Data frame including covariates}
}
\value{
A list of
\itemize{
  \item X_fe Design matrix for fixed effects
  \item X_re Design matrix for random effects
  \item S Smoothness matrix
  \item ncol_re Number of columns of X_re and S for each random effect
}
}
\description{
Create model matrices (observation model)
}
