% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{mvnorm_invlink}
\alias{mvnorm_invlink}
\title{Multivariate Normal inverse link function}
\usage{
mvnorm_invlink(x)
}
\arguments{
\item{x}{Vector of parameters on linear predictor scale (in the order:
means, SDs, correlations)}
}
\description{
Multivariate Normal inverse link function
}
