% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{mvnorm_link}
\alias{mvnorm_link}
\title{Multivariate Normal link function}
\usage{
mvnorm_link(x)
}
\arguments{
\item{x}{Vector of parameters on natural scale (in the order: means,
SDs, correlations)}
}
\description{
Multivariate Normal link function
}
