% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{prec_to_cov}
\alias{prec_to_cov}
\title{Get covariance matrix from precision matrix}
\usage{
prec_to_cov(prec_mat)
}
\arguments{
\item{prec_mat}{Precision matrix (either of 'matrix' type
or sparse matrix on which as.matrix can be used)}
}
\value{
Precision matrix
}
\description{
The covariance matrix is the inverse of the precision matrix. By default,
the function \code{solve} is used for inversion. If it fails (e.g.,
singular system), then \code{MASS::ginv} is used instead, and returns the
Moore-Penrose generalised inverse of the precision matrix.
}
