% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{quad_pos_solve}
\alias{quad_pos_solve}
\title{Solve for positive root of quadratic ax^2 + bx + c = 0 when it exists}
\usage{
quad_pos_solve(a, b, c)
}
\arguments{
\item{a}{coefficient of x^2}

\item{b}{coefficient of x}

\item{c}{scalar coefficient}
}
\value{
real positive root if it exists
}
\description{
Solve for positive root of quadratic ax^2 + bx + c = 0 when it exists
}
