% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_dist.R
\name{remove_dist}
\alias{remove_dist}
\title{Remove distribution from package's stored distributions}
\usage{
remove_dist(name, compile_cpp = TRUE)
}
\arguments{
\item{name}{name of distribution to remove}

\item{compile_cpp}{if FALSE then package is not recompiled after changes}
}
\description{
Remove distribution from package's stored distributions
}
