% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spde.R
\name{spde_loc}
\alias{spde_loc}
\title{Observation locations for an SPDE smooth}
\usage{
spde_loc(term, data)
}
\arguments{
\item{term}{Character vector of one or two covariate names}

\item{data}{Data frame containing those covariates}
}
\value{
Numeric vector (1d) or two-column matrix (2d) of locations
}
\description{
Observation locations for an SPDE smooth
}
