% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{strip_comments}
\alias{strip_comments}
\title{Strip comments marked with a hash from a character vector}
\usage{
strip_comments(str)
}
\arguments{
\item{str}{the character vector}
}
\value{
character vector with comments removed (and lines with only comments
completely removed)
}
\description{
Strip comments marked with a hash from a character vector
}
