% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{cov_grid}
\alias{cov_grid}
\title{Grid of covariates}
\usage{
cov_grid(var, data = NULL, obj = NULL, covs = NULL, formulas, n_grid = 1000)
}
\arguments{
\item{var}{Name of variable}

\item{data}{Data frame containing the covariates. If not provided, data
are extracted from obj}

\item{obj}{HMM model object containing data and formulas}

\item{covs}{Optional named list for values of covariates (other than 'var') 
that should be used in the plot (or dataframe with single row). If this is
not specified, the mean value is used for numeric variables, and the
first level for factor variables.}

\item{formulas}{List of formulas used in the model}

\item{n_grid}{Grid size (number of points). Default: 1000.}
}
\value{
Data frame of covariates, with 'var' defined over a grid,
and other covariates fixed to their mean (numeric) or first level
(factor).
}
\description{
Grid of covariates
}
