% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{Hmm}
\alias{Hmm}
\title{R6 class for hidden Markov model}
\description{
R6 class for hidden Markov model

R6 class for hidden Markov model
}
\details{
Encapsulates the observation and hidden state models for a hidden
Markov model.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Hmm$new()}}
\item \href{#method-obs}{\code{Hmm$obs()}}
\item \href{#method-hidden}{\code{Hmm$hidden()}}
\item \href{#method-res}{\code{Hmm$res()}}
\item \href{#method-tmb_obj}{\code{Hmm$tmb_obj()}}
\item \href{#method-tmb_rep}{\code{Hmm$tmb_rep()}}
\item \href{#method-states}{\code{Hmm$states()}}
\item \href{#method-par}{\code{Hmm$par()}}
\item \href{#method-nllk}{\code{Hmm$nllk()}}
\item \href{#method-setup}{\code{Hmm$setup()}}
\item \href{#method-fit}{\code{Hmm$fit()}}
\item \href{#method-viterbi}{\code{Hmm$viterbi()}}
\item \href{#method-CI_wpar}{\code{Hmm$CI_wpar()}}
\item \href{#method-plot_ts}{\code{Hmm$plot_ts()}}
\item \href{#method-clone}{\code{Hmm$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create new Hmm object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hmm$new(obs, hidden)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obs}}{Observation object}

\item{\code{hidden}}{MarkovChain object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new Hmm object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-obs"></a>}}
\subsection{Method \code{obs()}}{
Observation object for this model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hmm$obs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hidden"></a>}}
\subsection{Method \code{hidden()}}{
MarkovChain object for this model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hmm$hidden()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-res"></a>}}
\subsection{Method \code{res()}}{
Output of optimiser after model fitting
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hmm$res()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tmb_obj"></a>}}
\subsection{Method \code{tmb_obj()}}{
Model object created by TMB. This is the output of 
the TMB function \code{MakeADFun}, and it is a list including elements
\itemize{
  \item{\code{fn}}{Objective function}
  \item{\code{gr}}{Gradient function of fn}
  \item{\code{par}}{Vector of initial parameters on working scale}
}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hmm$tmb_obj()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tmb_rep"></a>}}
\subsection{Method \code{tmb_rep()}}{
Output of the TMB function \code{sdreport}, which includes 
estimates and standard errors for all model parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hmm$tmb_rep()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-states"></a>}}
\subsection{Method \code{states()}}{
Vector of estimated states, after \code{viterbi} has
been run
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hmm$states()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-par"></a>}}
\subsection{Method \code{par()}}{
Model parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hmm$par()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list with elements:
\itemize{
  \item{\code{obspar}}{Parameters of observation model}
  \item{\code{tpm}}{Transition probability matrix of hidden state model}
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nllk"></a>}}
\subsection{Method \code{nllk()}}{
Objective function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hmm$nllk(par)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{par}}{Vector of parameters for which the function should be
evaluated (on the working scale).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Negative log-likelihood
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setup"></a>}}
\subsection{Method \code{setup()}}{
TMB setup

This creates an attribute \code{tmb_obj}, which can be used to 
evaluate the negative log-likelihood function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hmm$setup()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\subsection{Method \code{fit()}}{
Model fitting

The negative log-likelihood of the model is minimised using the
function \code{optim}. TMB uses the Laplace approximation to integrate 
the random effects out of the likelihood.

After the model has been fitted, the output of \code{optim} can be
accessed using the method \code{res}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hmm$fit()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-viterbi"></a>}}
\subsection{Method \code{viterbi()}}{
Viterbi algorithm
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hmm$viterbi()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Most likely state sequence
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CI_wpar"></a>}}
\subsection{Method \code{CI_wpar()}}{
Confidence intervals for model parameters on the working scale 

These are Wald confidence intervals, obtained from the standard errors 
returned by the TMB function \code{sdreport}. See the TMB documentation 
for more details.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hmm$CI_wpar(level = 0.95)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{level}}{Confidence level (default: 0.95 for 95\% confidence intervals)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Matrix with three columns: (1) estimates, (2) lower bounds of
confidence intervals, (3) upper bounds of confidence intervals.
Time series plot coloured by states

Creates a plot of the data coloured by the most likely state sequence,
as estimated by the Viterbi algorithm. If one variable name is passed
as input, it is plotted against time. If two variables are passed, they
are plotted against each other.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_ts"></a>}}
\subsection{Method \code{plot_ts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hmm$plot_ts(var1, var2 = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{var1}}{Name of the variable to plot.}

\item{\code{var2}}{Optional name of a second variable, for 2-d plot.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hmm$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
