% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{Hmm}
\alias{Hmm}
\title{R6 class for hidden Markov model}
\description{
R6 class for hidden Markov model

R6 class for hidden Markov model
}
\details{
Encapsulates the observation and hidden state models for a hidden
Markov model.

This function transforms the smoothness parameter of
each smooth term into a standard deviation, given by 
SD = 1/sqrt(lambda). It is particularly helpful to get the
standard deviations of independent normal random effects.

This method generates confidence intervals by simulation.
That is, it generates \code{n_post} posterior samples of 
the estimated parameters from a multivariate normal distribution,
where the mean is the vector of estimates and the covariance matrix 
is provided by TMB. Then, transition probabilities are derived for 
each set of posterior parameter values, and confidence intervals
are obtained as quantiles of the posterior simulated transition
probabilities.


This method generates confidence intervals by simulation.
That is, it generates \code{n_post} posterior samples of 
the estimated parameters from a multivariate normal distribution,
where the mean is the vector of estimates and the covariance matrix 
is provided by TMB. Then, stationary state probabilities are derived 
for each set of posterior parameter values, and confidence intervals
are obtained as quantiles of the posterior simulated stationary state
probabilities.


This method generates confidence intervals by simulation.
That is, it generates \code{n_post} posterior samples of 
the estimated parameters from a multivariate normal distribution,
where the mean is the vector of estimates and the covariance matrix 
is provided by TMB. Then, observation parameters are derived for 
each set of posterior parameter values, and confidence intervals
are obtained as quantiles of the posterior simulated observation
parameters.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Hmm$new()}}
\item \href{#method-obs}{\code{Hmm$obs()}}
\item \href{#method-hidden}{\code{Hmm$hidden()}}
\item \href{#method-out}{\code{Hmm$out()}}
\item \href{#method-tmb_obj}{\code{Hmm$tmb_obj()}}
\item \href{#method-tmb_rep}{\code{Hmm$tmb_rep()}}
\item \href{#method-states}{\code{Hmm$states()}}
\item \href{#method-coeff_fe}{\code{Hmm$coeff_fe()}}
\item \href{#method-coeff_re}{\code{Hmm$coeff_re()}}
\item \href{#method-lambda}{\code{Hmm$lambda()}}
\item \href{#method-vcomp}{\code{Hmm$vcomp()}}
\item \href{#method-par}{\code{Hmm$par()}}
\item \href{#method-nllk}{\code{Hmm$nllk()}}
\item \href{#method-setup}{\code{Hmm$setup()}}
\item \href{#method-fit}{\code{Hmm$fit()}}
\item \href{#method-viterbi}{\code{Hmm$viterbi()}}
\item \href{#method-CI_coeff}{\code{Hmm$CI_coeff()}}
\item \href{#method-CI_tpm}{\code{Hmm$CI_tpm()}}
\item \href{#method-CI_statdist}{\code{Hmm$CI_statdist()}}
\item \href{#method-CI_obspar}{\code{Hmm$CI_obspar()}}
\item \href{#method-predict_tpm}{\code{Hmm$predict_tpm()}}
\item \href{#method-predict_obspar}{\code{Hmm$predict_obspar()}}
\item \href{#method-simulate}{\code{Hmm$simulate()}}
\item \href{#method-plot_ts}{\code{Hmm$plot_ts()}}
\item \href{#method-plot_tpm}{\code{Hmm$plot_tpm()}}
\item \href{#method-plot_stat_dist}{\code{Hmm$plot_stat_dist()}}
\item \href{#method-plot_obspar}{\code{Hmm$plot_obspar()}}
\item \href{#method-formulation}{\code{Hmm$formulation()}}
\item \href{#method-clone}{\code{Hmm$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create new Hmm object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hmm$new(obs, hidden)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obs}}{Observation object}

\item{\code{hidden}}{MarkovChain object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new Hmm object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-obs"></a>}}
\if{latex}{\out{\hypertarget{method-obs}{}}}
\subsection{Method \code{obs()}}{
Observation object for this model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hmm$obs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hidden"></a>}}
\if{latex}{\out{\hypertarget{method-hidden}{}}}
\subsection{Method \code{hidden()}}{
MarkovChain object for this model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hmm$hidden()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-out"></a>}}
\if{latex}{\out{\hypertarget{method-out}{}}}
\subsection{Method \code{out()}}{
Output of optimiser after model fitting
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hmm$out()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tmb_obj"></a>}}
\if{latex}{\out{\hypertarget{method-tmb_obj}{}}}
\subsection{Method \code{tmb_obj()}}{
Model object created by TMB. This is the output of 
the TMB function \code{MakeADFun}, and it is a list including elements
\itemize{
  \item{\code{fn}}{Objective function}
  \item{\code{gr}}{Gradient function of fn}
  \item{\code{par}}{Vector of initial parameters on working scale}
}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hmm$tmb_obj()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tmb_rep"></a>}}
\if{latex}{\out{\hypertarget{method-tmb_rep}{}}}
\subsection{Method \code{tmb_rep()}}{
Output of the TMB function \code{sdreport}, which includes 
estimates and standard errors for all model parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hmm$tmb_rep()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-states"></a>}}
\if{latex}{\out{\hypertarget{method-states}{}}}
\subsection{Method \code{states()}}{
Vector of estimated states, after \code{viterbi} has
been run
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hmm$states()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-coeff_fe"></a>}}
\if{latex}{\out{\hypertarget{method-coeff_fe}{}}}
\subsection{Method \code{coeff_fe()}}{
Coefficients for fixed effect parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hmm$coeff_fe()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-coeff_re"></a>}}
\if{latex}{\out{\hypertarget{method-coeff_re}{}}}
\subsection{Method \code{coeff_re()}}{
Coefficients for random effect parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hmm$coeff_re()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lambda"></a>}}
\if{latex}{\out{\hypertarget{method-lambda}{}}}
\subsection{Method \code{lambda()}}{
Smoothness parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hmm$lambda()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vcomp"></a>}}
\if{latex}{\out{\hypertarget{method-vcomp}{}}}
\subsection{Method \code{vcomp()}}{
Variance components of smooth terms
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hmm$vcomp()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-par"></a>}}
\if{latex}{\out{\hypertarget{method-par}{}}}
\subsection{Method \code{par()}}{
Model parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hmm$par()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list with elements:
\itemize{
  \item{\code{obspar}}{Parameters of observation model}
  \item{\code{tpm}}{Transition probability matrix of hidden state model}
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nllk"></a>}}
\if{latex}{\out{\hypertarget{method-nllk}{}}}
\subsection{Method \code{nllk()}}{
Objective function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hmm$nllk(par)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{par}}{Vector of parameters for which the function should be
evaluated (on the working scale).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Negative log-likelihood
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setup"></a>}}
\if{latex}{\out{\hypertarget{method-setup}{}}}
\subsection{Method \code{setup()}}{
TMB setup

This creates an attribute \code{tmb_obj}, which can be used to 
evaluate the negative log-likelihood function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hmm$setup(silent = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{silent}}{Logical. If TRUE, all tracing outputs are hidden (default).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\if{latex}{\out{\hypertarget{method-fit}{}}}
\subsection{Method \code{fit()}}{
Model fitting

The negative log-likelihood of the model is minimised using the
function \code{optim}. TMB uses the Laplace approximation to integrate 
the random effects out of the likelihood.

After the model has been fitted, the output of \code{optim} can be
accessed using the method \code{out}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hmm$fit(silent = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{silent}}{Logical. If TRUE, all tracing outputs are hidden (default).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-viterbi"></a>}}
\if{latex}{\out{\hypertarget{method-viterbi}{}}}
\subsection{Method \code{viterbi()}}{
Viterbi algorithm
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hmm$viterbi()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Most likely state sequence
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CI_coeff"></a>}}
\if{latex}{\out{\hypertarget{method-CI_coeff}{}}}
\subsection{Method \code{CI_coeff()}}{
Confidence intervals for model parameters on the working scale 

These are Wald confidence intervals, obtained from the standard errors 
returned by the TMB function \code{sdreport}. See the TMB documentation 
for more details.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hmm$CI_coeff(level = 0.95)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{level}}{Confidence level (default: 0.95 for 95\% confidence intervals)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Matrix with three columns: (1) estimates, (2) lower bounds of
confidence intervals, (3) upper bounds of confidence intervals.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CI_tpm"></a>}}
\if{latex}{\out{\hypertarget{method-CI_tpm}{}}}
\subsection{Method \code{CI_tpm()}}{
Confidence intervals for transition probabilities
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hmm$CI_tpm(X_fe, X_re, level = 0.95, n_post = 1000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X_fe}}{Design matrix for fixed effects, as returned
by \code{make_mat}}

\item{\code{X_re}}{Design matrix for random effects, as returned
by \code{make_mat}}

\item{\code{level}}{Confidence level (default: 0.95 for 95\% confidence 
intervals)}

\item{\code{n_post}}{Number of posterior samples from which the confidence
intervals are calculated. Larger values will reduce approximation
error, but increase computation time. Defaults to 1000.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List with elements:
\itemize{
  \item{\code{low}}{Array of lower bounds of confidence intervals.
  Each slice of the array is a transiton probability matrix.}
  \item{\code{upp}}{Array of upper bounds of confidence intervals.
  Each slice of the array is a transiton probability matrix.}
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CI_statdist"></a>}}
\if{latex}{\out{\hypertarget{method-CI_statdist}{}}}
\subsection{Method \code{CI_statdist()}}{
Confidence intervals for stationary distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hmm$CI_statdist(X_fe, X_re, level = 0.95, n_post = 1000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X_fe}}{Design matrix for fixed effects, as returned
by \code{make_mat}}

\item{\code{X_re}}{Design matrix for random effects, as returned
by \code{make_mat}}

\item{\code{level}}{Confidence level (default: 0.95 for 95\% confidence 
intervals)}

\item{\code{n_post}}{Number of posterior samples from which the confidence
intervals are calculated. Larger values will reduce approximation
error, but increase computation time. Defaults to 1000.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List with elements:
\itemize{
  \item{\code{low}}{Matrix of lower bounds of confidence intervals,
  with one row for each time step, and one column for each state.}
  \item{\code{upp}}{Matrix of upper bounds of confidence intervals,
  with one row for each time step, and one column for each state.}
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CI_obspar"></a>}}
\if{latex}{\out{\hypertarget{method-CI_obspar}{}}}
\subsection{Method \code{CI_obspar()}}{
Confidence intervals for observation parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hmm$CI_obspar(X_fe, X_re, level = 0.95, n_post = 1000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X_fe}}{Design matrix for fixed effects, as returned
by \code{make_mat}}

\item{\code{X_re}}{Design matrix for random effects, as returned
by \code{make_mat}}

\item{\code{level}}{Confidence level (default: 0.95 for 95\% confidence 
intervals)}

\item{\code{n_post}}{Number of posterior samples from which the confidence
intervals are calculated. Larger values will reduce approximation
error, but increase computation time. Defaults to 1000.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List with elements:
\itemize{
  \item{\code{low}}{Array of lower bounds of confidence intervals,
  with one slice for each time step, one row for each observation 
  parameter, and one column for each state.}
  \item{\code{upp}}{Array of upper bounds of confidence intervals,
  with one slice for each time step, one row for each observation 
  parameter, and one column for each state.}
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict_tpm"></a>}}
\if{latex}{\out{\hypertarget{method-predict_tpm}{}}}
\subsection{Method \code{predict_tpm()}}{
Predict transition probabilities
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hmm$predict_tpm(new_data = NULL, CI = FALSE, level = 0.95, n_post = 1000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_data}}{Data frame containing covariate values for which the
transition probabilities should be predicted.}

\item{\code{CI}}{Logical argument: should the function return confidence
intervals for the parameters? Default: FALSE.}

\item{\code{level}}{Confidence level (default: 0.95 for 95\% confidence 
intervals) if \code{CI = TRUE}.}

\item{\code{n_post}}{Number of posterior samples from which the confidence
intervals are calculated if \code{CI = TRUE}. Larger values will reduce 
approximation error, but increase computation time. Defaults to 1000.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
If \code{CI = FALSE}, returns an array of point estimates, 
where each slice is a transition probability matrix corresponding 
to one row of \code{new_data}. If \code{CI = TRUE}, returns a list 
with elements:
\itemize{
  \item{\code{estimate}}{Array of point estimates, where each slice is
  a transition probability matrix corresponding to one row of \code{new_data}.}
  \item{\code{low}}{Array of lower bounds of confidence intervals,
  with same structure as \code{estimate}.}
  \item{\code{upp}}{Array of upper bounds of confidence intervals,
  with same structure as \code{estimate}.}
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict_obspar"></a>}}
\if{latex}{\out{\hypertarget{method-predict_obspar}{}}}
\subsection{Method \code{predict_obspar()}}{
Predict observation parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hmm$predict_obspar(new_data = NULL, CI = FALSE, level = 0.95, n_post = 1000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_data}}{Data frame containing covariate values for which the
observation parameters should be predicted.}

\item{\code{CI}}{Logical argument: should the function return confidence
intervals for the parameters? Default: FALSE.}

\item{\code{level}}{Confidence level (default: 0.95 for 95\% confidence 
intervals) if \code{CI = TRUE}.}

\item{\code{n_post}}{Number of posterior samples from which the confidence
intervals are calculated if \code{CI = TRUE}. Larger values will reduce 
approximation error, but increase computation time. Defaults to 1000.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
If \code{CI = FALSE}, returns an array of point estimates, 
with one slice for each row of \code{new_data}, one row for each 
observation parameter, and one column for each state. If 
\code{CI = TRUE}, returns a list with elements:
\itemize{
  \item{\code{estimate}}{Array of point estimates, with one slice for 
  each row of \code{new_data}, one row for each observation 
  parameter, and one column for each state.}
  \item{\code{low}}{Array of lower bounds of confidence intervals,
  with same structure as \code{estimate}.}
  \item{\code{upp}}{Array of upper bounds of confidence intervals,
  with same structure as \code{estimate}.}
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-simulate"></a>}}
\if{latex}{\out{\hypertarget{method-simulate}{}}}
\subsection{Method \code{simulate()}}{
Simulate from hidden Markov model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hmm$simulate(n, data = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Number of time steps to simulate}

\item{\code{data}}{Optional data frame including covariates}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data frame including columns of data (if provided), and simulated
data variables
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_ts"></a>}}
\if{latex}{\out{\hypertarget{method-plot_ts}{}}}
\subsection{Method \code{plot_ts()}}{
Time series plot coloured by states

Creates a plot of the data coloured by the most likely state sequence,
as estimated by the Viterbi algorithm. If one variable name is passed
as input, it is plotted against time. If two variables are passed, they
are plotted against each other.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hmm$plot_ts(var1, var2 = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{var1}}{Name of the variable to plot.}

\item{\code{var2}}{Optional name of a second variable, for 2-d plot.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_tpm"></a>}}
\if{latex}{\out{\hypertarget{method-plot_tpm}{}}}
\subsection{Method \code{plot_tpm()}}{
Plot transition probability matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hmm$plot_tpm(var, covs = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{var}}{Name of covariate as a function of which the transition
probabilities should be plotted}

\item{\code{covs}}{Optional data frame with a single row and one column
for each covariate, giving the values that should be used. If this is
not specified, the mean value is used for numeric variables, and the
first level for factor variables.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_stat_dist"></a>}}
\if{latex}{\out{\hypertarget{method-plot_stat_dist}{}}}
\subsection{Method \code{plot_stat_dist()}}{
Plot stationary state probabilities
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hmm$plot_stat_dist(var, covs = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{var}}{Name of covariate as a function of which the state
probabilities should be plotted}

\item{\code{covs}}{Optional data frame with a single row and one column
for each covariate, giving the values that should be used. If this is
not specified, the mean value is used for numeric variables, and the
first level for factor variables.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_obspar"></a>}}
\if{latex}{\out{\hypertarget{method-plot_obspar}{}}}
\subsection{Method \code{plot_obspar()}}{
Plot observation parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hmm$plot_obspar(var, covs = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{var}}{Name of covariate as a function of which the parameters
should be plotted}

\item{\code{covs}}{Optional data frame with a single row and one column
for each covariate, giving the values that should be used. If this is
not specified, the mean value is used for numeric variables, and the
first level for factor variables.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-formulation"></a>}}
\if{latex}{\out{\hypertarget{method-formulation}{}}}
\subsection{Method \code{formulation()}}{
Print model formulation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hmm$formulation()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hmm$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
