% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmmTMB-package.R
\docType{package}
\name{hmmTMB-package}
\alias{hmmTMB}
\alias{hmmTMB-package}
\title{hmmTMB: Fit hidden Markov models using Template Model Builder}
\description{
Fitting hidden Markov models with Template Model Builder (TMB) using automatic differentiation and Laplace approximation, allowing for faster inference and more flexible random effect / smoothing spline structures.
}
\details{
The package hmmTMB is based on three main classes (i.e., types of objects):
MarkovChain, Observation, and HMM. Type ?MarkovChain, ?Observation or ?HMM
to find their documentation, or consult the package vignettes for detailed
examples of the hmmTMB workflow.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/TheoMichelot/hmmTMB}
}

}
\keyword{internal}
