% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmmTMB-package.R
\docType{package}
\name{hmmTMB-package}
\alias{hmmTMB}
\alias{hmmTMB-package}
\title{hmmTMB: Fit Hidden Markov Models using Template Model Builder}
\description{
Fitting hidden Markov models using automatic differentiation and Laplace approximation, allowing for fast inference and flexible covariate effects (including random effects and smoothing splines) on model parameters. Michelot (2022) \href{https://arxiv.org/abs/2211.14139}{arXiv:2211.14139}
}
\details{
The package hmmTMB is based on three main classes (i.e., types of objects):
MarkovChain, Observation, and HMM. Type ?MarkovChain, ?Observation or ?HMM
to find their documentation, or consult the package vignettes for detailed
examples of the hmmTMB workflow.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/TheoMichelot/hmmTMB}
}

}
\author{
\strong{Maintainer}: Theo Michelot \email{theo.michelot@dal.ca}

Authors:
\itemize{
  \item Richard Glennie [contributor]
}

}
\keyword{internal}
