% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden.R
\name{MarkovChain}
\alias{MarkovChain}
\title{R6 class for HMM hidden process model}
\description{
R6 class for HMM hidden process model

R6 class for HMM hidden process model
}
\details{
Contains the parameters and model formulas for the hidden process model.

This function transforms the smoothness parameter of
each smooth term into a standard deviation, given by 
SD = 1/sqrt(lambda). It is particularly helpful to get the
standard deviations of independent normal random effects.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MarkovChain$new()}}
\item \href{#method-structure}{\code{MarkovChain$structure()}}
\item \href{#method-formulas}{\code{MarkovChain$formulas()}}
\item \href{#method-tpm}{\code{MarkovChain$tpm()}}
\item \href{#method-coeff_fe}{\code{MarkovChain$coeff_fe()}}
\item \href{#method-coeff_re}{\code{MarkovChain$coeff_re()}}
\item \href{#method-lambda}{\code{MarkovChain$lambda()}}
\item \href{#method-vcomp}{\code{MarkovChain$vcomp()}}
\item \href{#method-nstates}{\code{MarkovChain$nstates()}}
\item \href{#method-terms}{\code{MarkovChain$terms()}}
\item \href{#method-update_tpm}{\code{MarkovChain$update_tpm()}}
\item \href{#method-update_coeff_fe}{\code{MarkovChain$update_coeff_fe()}}
\item \href{#method-update_coeff_re}{\code{MarkovChain$update_coeff_re()}}
\item \href{#method-update_lambda}{\code{MarkovChain$update_lambda()}}
\item \href{#method-make_mat}{\code{MarkovChain$make_mat()}}
\item \href{#method-make_mat_grid}{\code{MarkovChain$make_mat_grid()}}
\item \href{#method-tpm2par}{\code{MarkovChain$tpm2par()}}
\item \href{#method-par2tpm}{\code{MarkovChain$par2tpm()}}
\item \href{#method-tpm_all}{\code{MarkovChain$tpm_all()}}
\item \href{#method-stat_dists}{\code{MarkovChain$stat_dists()}}
\item \href{#method-simulate}{\code{MarkovChain$simulate()}}
\item \href{#method-formulation}{\code{MarkovChain$formulation()}}
\item \href{#method-clone}{\code{MarkovChain$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create new MarkovChain object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovChain$new(
  n_states = NULL,
  structure = NULL,
  tpm0 = NULL,
  coeff_fe0 = NULL,
  coeff_re0 = NULL,
  data = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_states}}{Number of states. If not specified, then \code{structure} 
needs to be provided as a matrix, and n_states is deduced from its dimensions.}

\item{\code{structure}}{Either (1) matrix with an entry of "." on diagonal, a "0" for 
transitions that are not allowed (not implemented yet), and a formula "~1" 
for covariates affecting transitions that are to be estimated, or (2) single
formula, assumed for all transition probabilities. (Default: no covariate
dependence.)}

\item{\code{tpm0}}{Initial transition probability matrix. (Default: 0.9 on diagonal,
and 0.1/(n_states - 1) for all other entries.) If the model has covariates,
then \code{tpm0} is used to set the intercept parameters for the transition
probabilities, and the other parameters are set to 0.}

\item{\code{coeff_fe0}}{Initial coefficients for fixed effects parameters. If
not provided, the intercepts are set using \code{tpm0}, and other 
coefficients are set to 0.}

\item{\code{coeff_re0}}{Initial coefficients for random effects parameters.
Defaults to 0 if not provided.}

\item{\code{data}}{HMMData object, needed if the model includes covariates}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new MarkovChain object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-structure"></a>}}
\if{latex}{\out{\hypertarget{method-structure}{}}}
\subsection{Method \code{structure()}}{
Structure of MarkovChain model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovChain$structure()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-formulas"></a>}}
\if{latex}{\out{\hypertarget{method-formulas}{}}}
\subsection{Method \code{formulas()}}{
List of formulas for MarkovChain model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovChain$formulas()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tpm"></a>}}
\if{latex}{\out{\hypertarget{method-tpm}{}}}
\subsection{Method \code{tpm()}}{
Current transition probability matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovChain$tpm()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-coeff_fe"></a>}}
\if{latex}{\out{\hypertarget{method-coeff_fe}{}}}
\subsection{Method \code{coeff_fe()}}{
Current parameter estimates (fixed effects)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovChain$coeff_fe()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-coeff_re"></a>}}
\if{latex}{\out{\hypertarget{method-coeff_re}{}}}
\subsection{Method \code{coeff_re()}}{
Current parameter estimates (random effects)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovChain$coeff_re()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lambda"></a>}}
\if{latex}{\out{\hypertarget{method-lambda}{}}}
\subsection{Method \code{lambda()}}{
Smoothness parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovChain$lambda()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vcomp"></a>}}
\if{latex}{\out{\hypertarget{method-vcomp}{}}}
\subsection{Method \code{vcomp()}}{
Variance components of smooth terms
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovChain$vcomp()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nstates"></a>}}
\if{latex}{\out{\hypertarget{method-nstates}{}}}
\subsection{Method \code{nstates()}}{
Number of states
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovChain$nstates()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-terms"></a>}}
\if{latex}{\out{\hypertarget{method-terms}{}}}
\subsection{Method \code{terms()}}{
Terms of model formulas
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovChain$terms()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_tpm"></a>}}
\if{latex}{\out{\hypertarget{method-update_tpm}{}}}
\subsection{Method \code{update_tpm()}}{
Update transition probability matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovChain$update_tpm(tpm)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tpm}}{New transition probability matrix}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_coeff_fe"></a>}}
\if{latex}{\out{\hypertarget{method-update_coeff_fe}{}}}
\subsection{Method \code{update_coeff_fe()}}{
Update coefficients for fixed effect parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovChain$update_coeff_fe(coeff_fe)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{coeff_fe}}{Vector of coefficients for fixed effect parameters}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_coeff_re"></a>}}
\if{latex}{\out{\hypertarget{method-update_coeff_re}{}}}
\subsection{Method \code{update_coeff_re()}}{
Update coefficients for random effect parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovChain$update_coeff_re(coeff_re)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{coeff_re}}{Vector of coefficients for random effect parameters}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_lambda"></a>}}
\if{latex}{\out{\hypertarget{method-update_lambda}{}}}
\subsection{Method \code{update_lambda()}}{
Update smoothness parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovChain$update_lambda(lambda)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lambda}}{New smoothness parameter vector}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_mat"></a>}}
\if{latex}{\out{\hypertarget{method-make_mat}{}}}
\subsection{Method \code{make_mat()}}{
Make model matrices
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovChain$make_mat(data, new_data = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data frame containing all needed covariates}

\item{\code{new_data}}{Optional new data set, including covariates for which
the design matrices should be created. This needs to be passed in addition
to the argument '\code{data}', for cases where smooth terms or factor
covariates are included, and the original data set is needed to determine
the full range of covariate values.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list with elements:
\itemize{
  \item{X_fe}{Design matrix for fixed effects}
  \item{X_re}{Design matrix for random effects}
  \item{S}{Smoothness matrix for random effects}
  \item{ncol_re}{Number of columns of X_re and S for each random effect}
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_mat_grid"></a>}}
\if{latex}{\out{\hypertarget{method-make_mat_grid}{}}}
\subsection{Method \code{make_mat_grid()}}{
Design matrices for grid of covariates

Used in plotting functions such as HMM$plot_tpm and HMM$plot_stat_dist
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovChain$make_mat_grid(var, data, covs = NULL, n_grid = 1000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{var}}{Name of variable}

\item{\code{data}}{Data frame containing the covariates}

\item{\code{covs}}{Optional data frame with a single row and one column
for each covariate, giving the values that should be used. If this is
not specified, the mean value is used for numeric variables, and the
first level for factor variables.}

\item{\code{n_grid}}{Grid size (number of points). Default: 1000.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list with the same elements as the output of make_mat, 
plus a data frame of covariates values.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tpm2par"></a>}}
\if{latex}{\out{\hypertarget{method-tpm2par}{}}}
\subsection{Method \code{tpm2par()}}{
Transform transition probabilities to working scale

Apply the multinomial logit link function to get the corresponding parameters on the
working scale (i.e., linear predictor scale).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovChain$tpm2par(tpm)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tpm}}{Transition probability matrix}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Vector of parameters on linear predictor scale
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-par2tpm"></a>}}
\if{latex}{\out{\hypertarget{method-par2tpm}{}}}
\subsection{Method \code{par2tpm()}}{
Transform working parameters to transition probabilities

Apply the inverse multinomial logit link function to transform the parameters on
the working scale (i.e., linear predictor scale) into the transition
probabilities.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovChain$par2tpm(par)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{par}}{Vector of parameters on working scale}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Transition probability matrix
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tpm_all"></a>}}
\if{latex}{\out{\hypertarget{method-tpm_all}{}}}
\subsection{Method \code{tpm_all()}}{
Get transition probability matrices from design matrices
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovChain$tpm_all(X_fe, X_re, coeff_fe = NULL, coeff_re = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X_fe}}{Design matrix for fixed effects, as returned
by \code{make_mat}}

\item{\code{X_re}}{Design matrix for random effects, as returned
by \code{make_mat}}

\item{\code{coeff_fe}}{Optional vector of coefficients for fixed effect
parameters. If this isn't provided, the model parameters
are used.}

\item{\code{coeff_re}}{Optional vector of coefficients for random 
effect parameters. If this isn't provided, the model parameters 
are used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Array with one slice for each transition probability matrix
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stat_dists"></a>}}
\if{latex}{\out{\hypertarget{method-stat_dists}{}}}
\subsection{Method \code{stat_dists()}}{
Stationary distributions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovChain$stat_dists(X_fe, X_re)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X_fe}}{Design matrix for fixed effects, as returned
by \code{make_mat}}

\item{\code{X_re}}{Design matrix for random effects, as returned
by \code{make_mat}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Matrix of stationary distributions. Each row corresponds to
a row of the design matrices, and each column corresponds to a state.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-simulate"></a>}}
\if{latex}{\out{\hypertarget{method-simulate}{}}}
\subsection{Method \code{simulate()}}{
Simulate from Markov chain
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovChain$simulate(n, data = NULL, new_data = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Number of time steps to simulate}

\item{\code{data}}{Optional data frame containing all needed covariates}

\item{\code{new_data}}{Optional new data set, including covariates for which
the design matrices should be created. This needs to be passed in addition
to the argument '\code{data}', for cases where smooth terms or factor
covariates are included, and the original data set is needed to determine
the full range of covariate values.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Sequence of states of simulated chain
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-formulation"></a>}}
\if{latex}{\out{\hypertarget{method-formulation}{}}}
\subsection{Method \code{formulation()}}{
Print model formulation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovChain$formulation()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkovChain$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
