% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation.R
\name{Observation}
\alias{Observation}
\title{R6 class for HMM observation model}
\description{
R6 class for HMM observation model

R6 class for HMM observation model
}
\details{
Contains the data, distributions, parameters, and formulas for
the observation model from a hidden Markov model.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Observation$new()}}
\item \href{#method-data}{\code{Observation$data()}}
\item \href{#method-dists}{\code{Observation$dists()}}
\item \href{#method-nstates}{\code{Observation$nstates()}}
\item \href{#method-par}{\code{Observation$par()}}
\item \href{#method-inipar}{\code{Observation$inipar()}}
\item \href{#method-coeff_fe}{\code{Observation$coeff_fe()}}
\item \href{#method-coeff_re}{\code{Observation$coeff_re()}}
\item \href{#method-X_fe}{\code{Observation$X_fe()}}
\item \href{#method-X_re}{\code{Observation$X_re()}}
\item \href{#method-lambda}{\code{Observation$lambda()}}
\item \href{#method-vcomp}{\code{Observation$vcomp()}}
\item \href{#method-formulas}{\code{Observation$formulas()}}
\item \href{#method-terms}{\code{Observation$terms()}}
\item \href{#method-obs_var}{\code{Observation$obs_var()}}
\item \href{#method-known_states}{\code{Observation$known_states()}}
\item \href{#method-update_par}{\code{Observation$update_par()}}
\item \href{#method-update_coeff_fe}{\code{Observation$update_coeff_fe()}}
\item \href{#method-update_coeff_re}{\code{Observation$update_coeff_re()}}
\item \href{#method-update_X_fe}{\code{Observation$update_X_fe()}}
\item \href{#method-update_X_re}{\code{Observation$update_X_re()}}
\item \href{#method-update_lambda}{\code{Observation$update_lambda()}}
\item \href{#method-update_data}{\code{Observation$update_data()}}
\item \href{#method-make_mat}{\code{Observation$make_mat()}}
\item \href{#method-make_newdata_grid}{\code{Observation$make_newdata_grid()}}
\item \href{#method-n2w}{\code{Observation$n2w()}}
\item \href{#method-w2n}{\code{Observation$w2n()}}
\item \href{#method-linpred}{\code{Observation$linpred()}}
\item \href{#method-obs_probs}{\code{Observation$obs_probs()}}
\item \href{#method-plot_dist}{\code{Observation$plot_dist()}}
\item \href{#method-formulation}{\code{Observation$formulation()}}
\item \href{#method-clone}{\code{Observation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create new Observation object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$new(data, dists, n_states, par = NULL, formulas = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data frame containing response variables and covariates}

\item{\code{dists}}{Named list of Distribution objects for each data stream}

\item{\code{n_states}}{Number of states (needed to construct model formulas)}

\item{\code{par}}{List of observation parameters (for covariate-free model)}

\item{\code{formulas}}{List of formulas for observation parameters}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new Observation object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data"></a>}}
\if{latex}{\out{\hypertarget{method-data}{}}}
\subsection{Method \code{data()}}{
Data frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dists"></a>}}
\if{latex}{\out{\hypertarget{method-dists}{}}}
\subsection{Method \code{dists()}}{
List of distributions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$dists()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nstates"></a>}}
\if{latex}{\out{\hypertarget{method-nstates}{}}}
\subsection{Method \code{nstates()}}{
Number of states
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$nstates()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-par"></a>}}
\if{latex}{\out{\hypertarget{method-par}{}}}
\subsection{Method \code{par()}}{
Parameters on natural scale
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$par(t = 1, full_names = TRUE, linpred = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t}}{time point, default t = 1; 
if "all" then return for all time points, otherwise return at time points given in t}

\item{\code{full_names}}{Logical. If TRUE, the rows of the output
are named in the format "variable.parameter" (default). If
FALSE, the rows are names in the format "parameter". The
latter is used in various internal functions, when the parameters
need to be passed on to an R function.}

\item{\code{linpred}}{custom linear predictor}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-inipar"></a>}}
\if{latex}{\out{\hypertarget{method-inipar}{}}}
\subsection{Method \code{inipar()}}{
Return initial parameter values supplied
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$inipar()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-coeff_fe"></a>}}
\if{latex}{\out{\hypertarget{method-coeff_fe}{}}}
\subsection{Method \code{coeff_fe()}}{
Fixed effect parameters on working scale
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$coeff_fe()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-coeff_re"></a>}}
\if{latex}{\out{\hypertarget{method-coeff_re}{}}}
\subsection{Method \code{coeff_re()}}{
Random effect parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$coeff_re()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-X_fe"></a>}}
\if{latex}{\out{\hypertarget{method-X_fe}{}}}
\subsection{Method \code{X_fe()}}{
Fixed effect design matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$X_fe()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-X_re"></a>}}
\if{latex}{\out{\hypertarget{method-X_re}{}}}
\subsection{Method \code{X_re()}}{
Random effect design matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$X_re()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lambda"></a>}}
\if{latex}{\out{\hypertarget{method-lambda}{}}}
\subsection{Method \code{lambda()}}{
Smoothness parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$lambda()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vcomp"></a>}}
\if{latex}{\out{\hypertarget{method-vcomp}{}}}
\subsection{Method \code{vcomp()}}{
Variance components of smooth terms

This function transforms the smoothness parameter of
each smooth term into a standard deviation, given by 
SD = 1/sqrt(lambda). It is particularly helpful to get the
standard deviations of independent normal random effects.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$vcomp()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-formulas"></a>}}
\if{latex}{\out{\hypertarget{method-formulas}{}}}
\subsection{Method \code{formulas()}}{
List of model formulas for observation model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$formulas(raw = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-terms"></a>}}
\if{latex}{\out{\hypertarget{method-terms}{}}}
\subsection{Method \code{terms()}}{
Terms of model formulas
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$terms()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-obs_var"></a>}}
\if{latex}{\out{\hypertarget{method-obs_var}{}}}
\subsection{Method \code{obs_var()}}{
Data frame of response variables
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$obs_var(expand = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{expand}}{if TRUE then multivariate variables in observations are expanded 
to be univariate, creating extra columns}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-known_states"></a>}}
\if{latex}{\out{\hypertarget{method-known_states}{}}}
\subsection{Method \code{known_states()}}{
Vector of known states
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$known_states()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_par"></a>}}
\if{latex}{\out{\hypertarget{method-update_par}{}}}
\subsection{Method \code{update_par()}}{
Update parameters

Updates the 'par' attribute to the list passed as input,
and updates the intercept elements of 'coeff_fe' using
the list passed as input
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$update_par(par)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{par}}{New list of parameters}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_coeff_fe"></a>}}
\if{latex}{\out{\hypertarget{method-update_coeff_fe}{}}}
\subsection{Method \code{update_coeff_fe()}}{
Update coefficients for fixed effect parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$update_coeff_fe(coeff_fe)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{coeff_fe}}{New vector of coefficients for fixed effect 
parameters}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_coeff_re"></a>}}
\if{latex}{\out{\hypertarget{method-update_coeff_re}{}}}
\subsection{Method \code{update_coeff_re()}}{
Update random effect parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$update_coeff_re(coeff_re)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{coeff_re}}{New vector of coefficients for random effect 
parameters}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_X_fe"></a>}}
\if{latex}{\out{\hypertarget{method-update_X_fe}{}}}
\subsection{Method \code{update_X_fe()}}{
Update fixed effect design matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$update_X_fe(X_fe)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X_fe}}{New fixed effect design matrix}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_X_re"></a>}}
\if{latex}{\out{\hypertarget{method-update_X_re}{}}}
\subsection{Method \code{update_X_re()}}{
Update random effect design matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$update_X_re(X_re)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X_fe}}{New random effect design matrix}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_lambda"></a>}}
\if{latex}{\out{\hypertarget{method-update_lambda}{}}}
\subsection{Method \code{update_lambda()}}{
Update smoothness parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$update_lambda(lambda)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lambda}}{New smoothness parameter vector}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_data"></a>}}
\if{latex}{\out{\hypertarget{method-update_data}{}}}
\subsection{Method \code{update_data()}}{
Update data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$update_data(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{New data frame}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_mat"></a>}}
\if{latex}{\out{\hypertarget{method-make_mat}{}}}
\subsection{Method \code{make_mat()}}{
Make model matrices
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$make_mat(new_data = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_data}}{Optional new data set, including covariates for which
the design matrices should be created. If this argument is not specified,
the design matrices are based on the original data frame.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list with elements:
\itemize{
  \item{X_fe}{Design matrix for fixed effects}
  \item{X_re}{Design matrix for random effects}
  \item{S}{Smoothness matrix for random effects}
  \item{ncol_fe}{Number of columns of X_fe for each parameter}
  \item{ncol_re}{Number of columns of X_re and S for each random effect}
}
Design matrices for grid of covariates
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_newdata_grid"></a>}}
\if{latex}{\out{\hypertarget{method-make_newdata_grid}{}}}
\subsection{Method \code{make_newdata_grid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$make_newdata_grid(var, covs = NULL, n_grid = 1000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{var}}{Name of variable}

\item{\code{covs}}{Optional data frame with a single row and one column
for each covariate, giving the values that should be used. If this is
not specified, the mean value is used for numeric variables, and the
first level for factor variables.}

\item{\code{n_grid}}{Grid size (number of points). Default: 1000.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list with the same elements as the output of make_mat, 
plus a data frame of covariates values.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-n2w"></a>}}
\if{latex}{\out{\hypertarget{method-n2w}{}}}
\subsection{Method \code{n2w()}}{
Natural to working parameter transformation

This function applies the link functions of the distribution
parameters, to transform parameters from their natural scale
to the working scale (i.e., linear predictor scale)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$n2w(par)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{par}}{List of parameters on natural scale}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Vector of parameters on working scale
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-w2n"></a>}}
\if{latex}{\out{\hypertarget{method-w2n}{}}}
\subsection{Method \code{w2n()}}{
Working to natural parameter transformation

This function applies the inverse link functions of the
distribution parameters, to transform parameters from the working
scale (i.e., linear predictor scale) to their natural scale.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$w2n(coeff_fe)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{coeff_fe}}{Vector of parameters on working scale}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List of parameters on natural scale
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-linpred"></a>}}
\if{latex}{\out{\hypertarget{method-linpred}{}}}
\subsection{Method \code{linpred()}}{
Compute linear predictor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$linpred()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-obs_probs"></a>}}
\if{latex}{\out{\hypertarget{method-obs_probs}{}}}
\subsection{Method \code{obs_probs()}}{
Observation likelihoods
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$obs_probs(data = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{optional dataframe to include in form of obs_var() output}

\item{\code{X_fe}}{Design matrix for fixed effects}

\item{\code{X_re}}{Design matrix for random effects}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Matrix of likelihoods of observations, with one row for each 
time step, and one column for each state.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_dist"></a>}}
\if{latex}{\out{\hypertarget{method-plot_dist}{}}}
\subsection{Method \code{plot_dist()}}{
Plot histogram of data and pdfs

Plot histogram of observations for the variable specified by the argument name, 
overlaid with the pdf of the specified distribution for that data stream. 
Helpful to select initial parameter values for model fitting, or to visualise 
fitted state-dependent distributions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$plot_dist(name, weights = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of response variable for which the histogram
and pdfs should be plotted.}

\item{\code{weights}}{Optional vector of length the number of pdfs that are
plotted. Useful to visualise a mixture of distributions weighted by the
proportion of time spent in the different states.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-formulation"></a>}}
\if{latex}{\out{\hypertarget{method-formulation}{}}}
\subsection{Method \code{formulation()}}{
Print model formulation
Check constructor arguments 
Create a distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$formulation()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
