% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation.R
\name{Observation}
\alias{Observation}
\title{R6 class for HMM observation model}
\description{
R6 class for HMM observation model

R6 class for HMM observation model
}
\details{
Contains the data, distributions, parameters, and formulas for
the observation model from a hidden Markov model.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Observation$new()}}
\item \href{#method-data}{\code{Observation$data()}}
\item \href{#method-dists}{\code{Observation$dists()}}
\item \href{#method-nstates}{\code{Observation$nstates()}}
\item \href{#method-par}{\code{Observation$par()}}
\item \href{#method-wpar}{\code{Observation$wpar()}}
\item \href{#method-wpar_re}{\code{Observation$wpar_re()}}
\item \href{#method-formulas}{\code{Observation$formulas()}}
\item \href{#method-obs_var}{\code{Observation$obs_var()}}
\item \href{#method-update_par}{\code{Observation$update_par()}}
\item \href{#method-update_wpar}{\code{Observation$update_wpar()}}
\item \href{#method-update_wpar_re}{\code{Observation$update_wpar_re()}}
\item \href{#method-make_mat}{\code{Observation$make_mat()}}
\item \href{#method-make_mat_grid}{\code{Observation$make_mat_grid()}}
\item \href{#method-par_all}{\code{Observation$par_all()}}
\item \href{#method-obs_probs}{\code{Observation$obs_probs()}}
\item \href{#method-n2w}{\code{Observation$n2w()}}
\item \href{#method-w2n}{\code{Observation$w2n()}}
\item \href{#method-plot_dist}{\code{Observation$plot_dist()}}
\item \href{#method-clone}{\code{Observation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create new Observation object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$new(
  data,
  dists,
  n_states,
  par = NULL,
  wpar = NULL,
  wpar_re = NULL,
  formulas = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{HmmData object}

\item{\code{dists}}{Named list of Distribution objects for each data stream}

\item{\code{n_states}}{Number of states (needed to construct model formulas)}

\item{\code{par}}{List of observation parameters (for covariate-free model)}

\item{\code{wpar}}{Vector of fixed effect parameters on working scale}

\item{\code{wpar_re}}{Vector of random effect parameters. Defaults to a
vector of zeros if not provided.}

\item{\code{formulas}}{List of formulas for observation parameters}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new Observation object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data"></a>}}
\subsection{Method \code{data()}}{
HmmData object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dists"></a>}}
\subsection{Method \code{dists()}}{
List of distributions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$dists()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nstates"></a>}}
\subsection{Method \code{nstates()}}{
Number of states
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$nstates()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-par"></a>}}
\subsection{Method \code{par()}}{
Parameters on natural scale
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$par()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wpar"></a>}}
\subsection{Method \code{wpar()}}{
Fixed effect parameters on working scale
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$wpar()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wpar_re"></a>}}
\subsection{Method \code{wpar_re()}}{
Random effect parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$wpar_re()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-formulas"></a>}}
\subsection{Method \code{formulas()}}{
List of model formulas for observation model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$formulas()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-obs_var"></a>}}
\subsection{Method \code{obs_var()}}{
Data frame of response variables
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$obs_var()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_par"></a>}}
\subsection{Method \code{update_par()}}{
Update parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$update_par(par)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{par}}{New list of parameters}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_wpar"></a>}}
\subsection{Method \code{update_wpar()}}{
Update fixed effect parameters on working scale
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$update_wpar(wpar)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{wpar}}{New vector of fixed effect parameters on working scale}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_wpar_re"></a>}}
\subsection{Method \code{update_wpar_re()}}{
Update random effect parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$update_wpar_re(wpar_re)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{wpar_re}}{New vector og random effect parameters}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_mat"></a>}}
\subsection{Method \code{make_mat()}}{
Make model matrices
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$make_mat(new_data = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_data}}{Optional new data set, including covariates for which
the design matrices should be created. If this argument is not specified,
the design matrices are based on the original data frame.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list with elements:
\itemize{
  \item{X_fe}{Design matrix for fixed effects}
  \item{X_re}{Design matrix for random effects}
  \item{S}{Smoothness matrix for random effects}
  \item{ncol_re}{Number of columns of X_re and S for each random effect}
}
Design matrices for grid of covariates
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_mat_grid"></a>}}
\subsection{Method \code{make_mat_grid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$make_mat_grid(var, covs = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{var}}{Name of variable}

\item{\code{covs}}{Optional data frame with a single row and one column
for each covariate, giving the values that should be used. If this is
not specified, the mean value is used for numeric variables, and the
first level for factor variables.}

\item{\code{data}}{Data frame containing the covariates}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list with the same elements as the output of make_mat, 
plus a data frame of covariates values.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-par_all"></a>}}
\subsection{Method \code{par_all()}}{
Get observation parameters from design matrices
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$par_all(X_fe, X_re)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X_fe}}{Design matrix for fixed effects, as returned
by \code{make_mat}}

\item{\code{X_re}}{Design matrix for random effects, as returned
by \code{make_mat}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Array with one slice for each transition probability matrix
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-obs_probs"></a>}}
\subsection{Method \code{obs_probs()}}{
Observation likelihoods
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$obs_probs(X_fe, X_re)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X_fe}}{Design matrix for fixed effects}

\item{\code{X_re}}{Design matrix for random effects}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Matrix of likelihoods of observations, with one row for each 
time step, and one column for each state.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-n2w"></a>}}
\subsection{Method \code{n2w()}}{
Natural to working parameter transformation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$n2w(par)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{par}}{List of parameters on natural scale}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Vector of parameters on working scale
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-w2n"></a>}}
\subsection{Method \code{w2n()}}{
Working to natural parameter transformation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$w2n(wpar)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{wpar}}{Vector of parameters on working scale}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List of parameters on natural scale
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_dist"></a>}}
\subsection{Method \code{plot_dist()}}{
Plot histogram of data and pdfs

Plot histogram of observations for the variable specified by the argument name, 
overlaid with the pdf of the specified distribution for that data stream. 
Helpful to select initial parameter values for model fitting, or to visualise 
fitted state-dependent distributions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$plot_dist(name, par = NULL, weights = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of response variable for which the histogram
and pdfs should be plotted.}

\item{\code{par}}{Optional matrix of parameters, with one row for each state
and one column for each parameter. The columns of the matrix should be
named with the names of the parameters, e.g. "mean" and "sd" for normal
distribution. If not provided, the parameters stored in the object are
used (default).}

\item{\code{weights}}{Optional vector of length the number of pdfs that are
plotted. Useful to visualise a mixture of distributions weighted by the
proportion of time spent in the different states.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Observation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
